[RECIPE-PROJECT] Added Edit profile feature
