+++ recipe-hooks-project/src/Profile.js
@@ -1,79 +1,26 @@
 import React, {memo, useEffect} from "react";
-import styled from "styled-components";
 import axios from "axios";
 import "@fortawesome/fontawesome-free/css/all.css"
 import {Wrapper} from "./components/General";
+import {ProfileCard, Section, Title, Label, Input, EditSection, ActionIcon} from "./components/Profile";
 import {RECIPE_API_BASE_URL} from "./App";
 import useInputState from "./hooks/useInputState";
 import useToggle from "./hooks/useToggleState";
 
-const ProfileCard = styled.form`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  width: 30%;
-  border: 2px solid lightgray;
-  box-shadow: 0 19px 38px rgba(0, 0, 0, 0.3), 0 15px 12px rgba(0, 0, 0, 0.1);
-  margin: auto;
-`;
-const Section = styled.div`
-  display: flex;
-  justify-content: left;
-  align-items: center;
-  width: 90%;
-  padding-bottom: 1.5em;
-  margin: auto;
-`;
-const Title = styled(Section)`
-  height: 1em;
-  font-size: 1.5em;
-  font-weight: 500;
-  margin: auto;
-  justify-content: center;
-  padding-top: 1.5em;
-`;
-const Label = styled.label`
-  font-size: 1em;
-  width: 20%;
-  color: blue;
-  margin-left: 2.5em;
-`;
-const Input = styled.input.attrs((props) => ({
-  ...props
-}))`
-  font-size: 1em;
-  padding: 15px;
-  width: 50%;
-  border: 2px solid blue;
-  border-radius: 5px;
-  
-  &:hover {
-    border-color: red;
-  }
-`;
-const EditSection = styled(Section)`
-  margin: auto;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  cursor: pointer;
-`;
+const USER_PROFILE_API_RELATIVE_URL = "/user/me/";
 
 function Profile() {
   const [isEditing, toggleIsEditing] = useToggle(false);
-  const [email, handleEmailChange, setEmail] = useInputState("");
+  const [email, , setEmail] = useInputState("");
   const [password, handlePasswordChange] = useInputState("");
   const [name, handleNameChange, setName] = useInputState("");
 
   useEffect(() => {
     async function fetchData() {
-      const profile = await axios.get(RECIPE_API_BASE_URL + "/user/me",
+      const profile = await axios.get(RECIPE_API_BASE_URL + USER_PROFILE_API_RELATIVE_URL,
         {headers: {Authorization: "Token 771588d4be688173e35ffe08caec07ac8a95009e"}});
       setEmail(profile.data.email);
       setName(profile.data.name);
-
-      console.log("RERENDER")
     }
 
     fetchData();
@@ -82,6 +29,24 @@ function Profile() {
   function toggleReadonly(e) {
     e.preventDefault();
     toggleIsEditing();
+  }
+
+  function handleSubmit(e) {
+    e.preventDefault();
+    if (modifyUser(password, name))
+      toggleIsEditing();
+    else
+      console.log("Error"); //FIXME RETURN MESSAGE?
+  }
+
+  async function modifyUser(password, name) {
+    let modifyUserPayload;
+    if (password !== "")
+      modifyUserPayload = {name: name, password: password}
+    else
+      modifyUserPayload = {name: name}
+    const response = await axios.patch(RECIPE_API_BASE_URL + USER_PROFILE_API_RELATIVE_URL, modifyUserPayload);
+    return response.status === 200;
   }
 
   return (
@@ -96,7 +61,7 @@ function Profile() {
         </Section>
         <Section>
           <Label htmlFor="emailInput">Email: </Label>
-          <Input type="text" id="emailInput" value={email} onChange={handleEmailChange} readOnly={!isEditing}/>
+          <Input type="text" id="emailInput" value={email} readOnly={true}/>
         </Section>
         <Section>
           <Label htmlFor="passwordInput">Password: </Label>
@@ -104,7 +69,10 @@ function Profile() {
                  onChange={handlePasswordChange} readOnly={!isEditing}/>
         </Section>
         <EditSection>
-          <i className="fas fa-edit fa-2x" onClick={toggleReadonly}></i>
+          {isEditing
+            ? <ActionIcon className="fas fa-save fa-2x" onClick={handleSubmit}/>
+            : <ActionIcon className="fas fa-edit fa-2x" onClick={toggleReadonly}/>
+          }
         </EditSection>
       </ProfileCard>
     </Wrapper>
